+++ sharding-core/sharding-core-execute/src/main/java/org/apache/shardingsphere/sharding/execute/sql/execute/SQLExecuteCallback.java
@@ -20,10 +20,10 @@ package org.apache.shardingsphere.sharding.execute.sql.execute;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.sharding.execute.sql.StatementExecuteUnit;
 import org.apache.shardingsphere.sharding.execute.sql.execute.threadlocal.ExecutorExceptionHandler;
-import org.apache.shardingsphere.sharding.execute.sql.hook.SPISQLExecutionHook;
+import org.apache.shardingsphere.underlying.executor.hook.SPISQLExecutionHook;
 import org.apache.shardingsphere.underlying.executor.constant.ConnectionMode;
 import org.apache.shardingsphere.underlying.executor.engine.GroupedCallback;
-import org.apache.shardingsphere.sharding.execute.sql.hook.SQLExecutionHook;
+import org.apache.shardingsphere.underlying.executor.hook.SQLExecutionHook;
 import org.apache.shardingsphere.underlying.executor.context.ExecutionUnit;
 import org.apache.shardingsphere.spi.database.metadata.DataSourceMetaData;
 import org.apache.shardingsphere.spi.database.type.DatabaseType;
