+++ shardingsphere-underlying/shardingsphere-executor/src/main/java/org/apache/shardingsphere/underlying/executor/hook/SPISQLExecutionHook.java
@@ -15,50 +15,45 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.execute.sql.hook.fixture;
+package org.apache.shardingsphere.underlying.executor.hook;
 
-import org.apache.shardingsphere.sharding.execute.sql.hook.SQLExecutionHook;
+import org.apache.shardingsphere.spi.NewInstanceServiceLoader;
 import org.apache.shardingsphere.spi.database.metadata.DataSourceMetaData;
 
 import java.util.Collection;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
-public final class SQLExecutionHookFixture implements SQLExecutionHook {
+/**
+ * SQL Execution hook for SPI.
+ */
+public final class SPISQLExecutionHook implements SQLExecutionHook {
     
-    private static final Collection<String> ACTIONS = new LinkedList<>();
+    private final Collection<SQLExecutionHook> sqlExecutionHooks = NewInstanceServiceLoader.newServiceInstances(SQLExecutionHook.class);
+    
+    static {
+        NewInstanceServiceLoader.register(SQLExecutionHook.class);
+    }
     
     @Override
     public void start(final String dataSourceName, final String sql, final List<Object> parameters, 
                       final DataSourceMetaData dataSourceMetaData, final boolean isTrunkThread, final Map<String, Object> shardingExecuteDataMap) {
-        ACTIONS.add("start");
+        for (SQLExecutionHook each : sqlExecutionHooks) {
+            each.start(dataSourceName, sql, parameters, dataSourceMetaData, isTrunkThread, shardingExecuteDataMap);
+        }
     }
     
     @Override
     public void finishSuccess() {
-        ACTIONS.add("finishSuccess");
+        for (SQLExecutionHook each : sqlExecutionHooks) {
+            each.finishSuccess();
+        }
     }
     
     @Override
     public void finishFailure(final Exception cause) {
-        ACTIONS.add("finishFailure");
-    }
-    
-    /**
-     * Contains action or not.
-     * 
-     * @param action action
-     * @return contains action or not
-     */
-    public static boolean containsAction(final String action) {
-        return ACTIONS.contains(action);
-    }
-    
-    /**
-     * Clear actions.
-     */
-    public static void clearActions() {
-        ACTIONS.clear();
+        for (SQLExecutionHook each : sqlExecutionHooks) {
+            each.finishFailure(cause);
+        }
     }
 }
