+++ shardingsphere-underlying/shardingsphere-executor/src/test/resources/META-INF/services/org.apache.shardingsphere.underlying.executor.hook.SQLExecutionHook
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.sharding.execute.sql.hook.fixture.SQLExecutionHookFixture
+org.apache.shardingsphere.underlying.executor.hook.fixture.SQLExecutionHookFixture
